requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
python-dotenv
advanced-alchemy
asyncpg